refactor: runs smooth but doesn't delete all edges

+++ DataStructers/Graph.cpp
@@ -2,13 +2,16 @@
 
 Graph::Graph(Graph &G){
 	
-	this->_size = G._size;
-	this->_adj_arr = new List[this->_size];
-	this->_t = G._t;
-	this->_s = G._s;
-	
-	for(int i = 1; i < this->_size; i++){
-		this->_adj_arr[i].copyList(G._adj_arr[i]);
+	if(G.getSize() != 0){
+		this->_size = G._size;
+		this->_adj_arr = new List[this->_size];
+		this->_t = G._t;
+		this->_s = G._s;
+		
+		for(int i = 1; i < this->_size; i++){
+			if(!G.getAdjList(i).isEmpty())
+				this->_adj_arr[i].copyList(G._adj_arr[i]);
+		}
 	}
 }
 Graph::~Graph(){
@@ -18,13 +21,12 @@ Graph::~Graph(){
 }
 void Graph::makeEmptyGraph(int n){
 	
-	this->_size = n + 1;
-	
 	if(this->_size != 0){
 		this->deleteGraph();
 		delete[] this->_adj_arr;
 	}
 	
+	this->_size = n + 1;
 	this->_adj_arr = new List[this->_size];
 }
 void Graph::deleteGraph(){// This func deletes all the graph'_s adj lists
@@ -33,7 +35,6 @@ void Graph::deleteGraph(){// This func deletes all the graph'_s adj lists
 		
 		for(int i = 1; i < this->_size; i++){//The loop runs from 1 to n
 			this->_adj_arr[i].emptyList();
-			cout << i << " ";
 		}
 	}
 }
@@ -43,7 +44,7 @@ bool Graph::isAdj(int u, int v){
 }
 int Graph::addEdge(int u, int v){
 	
-	this->_adj_arr[u].insertToHead(new Vertex(v));
+	this->_adj_arr[u].insertToHead(v);
 	
 	if(this->_adj_arr[u].find(v)){
 		return 1;
@@ -52,9 +53,12 @@ int Graph::addEdge(int u, int v){
 		return 0;
 	}
 }
-void Graph::removeEdge(int u, int v){
+Vertex *Graph::removeEdge(int u, int v){
 	
-	this->_adj_arr[u].deleteNode(v);
+	if(!this->_adj_arr[u].isEmpty())
+		return this->_adj_arr[u].deleteNode(v);
+	
+	return nullptr;
 }
 void Graph::printGraph(){
 	
@@ -65,10 +69,10 @@ void Graph::printGraph(){
 }
 void Graph::readGraph(){
 	
-	int u, v, counter = 0;
+	int u, v;
 	string f_name;
 	fstream fp;
-	
+	//TODO: change to stdin input
 	if(this->_size != 0){//Checking whether this graph is already allocated
 		this->deleteGraph();
 		delete[] this->_adj_arr;
@@ -93,7 +97,7 @@ void Graph::readGraph(){
 		fp >> u;
 		fp >> v;
 		
-		this->_adj_arr[u].insertToHead(new Vertex(v));
+		this->_adj_arr[u].insertToHead(v);
 	}
 	
 	fp.close();
@@ -114,7 +118,7 @@ void Graph::makeGt(Graph &Gt){//Creating Gt graph
 		Vertex *curr = this->_adj_arr[i].getHead();
 		
 		while(curr != nullptr){
-			Gt._adj_arr[curr->data()].insertToHead(new Vertex(i));// inserting the opposite edge to Gt
+			Gt._adj_arr[curr->data()].insertToHead(i);// inserting the opposite edge to Gt
 			
 			curr = curr->getNext();
 		}
